[修改] 优化

+++ src/main/java/com/ruoyi/framework/shiro/service/PasswordService.java
@@ -50,9 +50,9 @@ public class PasswordService {
             retryCount = new AtomicInteger(0);
             loginRecordCache.put(loginName, retryCount);
         }
-        if (retryCount.incrementAndGet() > Integer.valueOf(maxRetryCount).intValue()) {
+        if (retryCount.incrementAndGet() > Integer.valueOf(maxRetryCount)) {
             AsyncManager.me().execute(AsyncFactory.recordLogininfor(loginName, Constants.LOGIN_FAIL, MessageUtils.message("user.password.retry.limit.exceed", maxRetryCount)));
-            throw new UserPasswordRetryLimitExceedException(Integer.valueOf(maxRetryCount).intValue());
+            throw new UserPasswordRetryLimitExceedException(Integer.valueOf(maxRetryCount));
         }
 
         if (!matches(user, password)) {
